pyproject.toml
[project]
name = "cheatsheet-generator"
version = "0.1.0"
description = "A tool to generate cheatsheets from topics and subtopics using CLAUDE"
dependencies = [
    "anthropic>=0.59.0",
    "pyyaml>=6.0.2",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[project.optional-dependencies]
test = [
    "pytest>=8.4.1",
    "pytest-cov",
    "pytest-mock",
    "pytest-html",
]

[tool.pytest.ini_options]
markers =[
    "api_test: quick check for API test",
]
testpaths=["tests"]
pythonpath=["."]
addopts = "-v --disable-warnings --cov=. --cov-report=html"

[tool.tox]
requires = ["tox>=4.19"]
env_list = ["3.11", "3.12"]

[tool.tox.env_run_base]
description = "Run test under {base_python}"
commands = [["pytest"]]

[tool.tox.env.type]
description = "run type check on code base"
deps = ["mypy==1.11.2", "types-cachetools>=5.5.0.20240820", "types-chardet>=5.0.4.6"]
commands = [["mypy", "src{/}tox"], ["mypy", "tests"]]